Removed the use of 'Tenant' at the moment to better consider its role and applicability in the domain.
