+++ ecomarkets/src/main/java/ecomarkets/domain/core/product/ImageRepository.java
@@ -1,0 +1,19 @@
+package ecomarkets.domain.core.product;
+
+import jakarta.annotation.Resource;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import org.eclipse.microprofile.config.inject.ConfigProperty;
+import software.amazon.awssdk.core.sync.RequestBody;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.GetObjectRequest;
+import software.amazon.awssdk.services.s3.model.PutObjectRequest;
+import software.amazon.awssdk.services.s3.model.PutObjectResponse;
+
+import java.nio.file.Path;
+
+public interface ImageRepository {
+
+    public void save(Path file);
+
+}
