+++ ecomarkets/src/main/resources/application.properties
@@ -5,3 +5,5 @@ quarkus.resteasy-reactive.path=/api
 
 quarkus.datasource.db-kind=mysql
 quarkus.hibernate-orm.database.generation=update
+
+bucket.name=ecomarkets
