+++ ecomarkets/src/test/java/ecomarkets/domain/core/product/ProductImageTest.java
@@ -1,69 +1,36 @@
 package ecomarkets.domain.core.product;
 
-import io.quarkus.test.TestTransaction;
 import io.quarkus.test.junit.QuarkusTest;
+import jakarta.annotation.Resource;
+import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
-import org.junit.Rule;
+import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
-import org.testcontainers.containers.localstack.LocalStackContainer;
-import org.testcontainers.utility.DockerImageName;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.CreateBucketRequest;
 import software.amazon.awssdk.services.s3.model.ListObjectsRequest;
-import software.amazon.awssdk.services.s3.model.ListObjectsResponse;
 import software.amazon.awssdk.services.s3.model.S3Object;
+import org.eclipse.microprofile.config.inject.ConfigProperty;
 
-import javax.tools.FileObject;
 import java.io.File;
-import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.is;
-import static org.testcontainers.containers.localstack.LocalStackContainer.Service.S3;
 
 @QuarkusTest
 public class ProductImageTest {
-    static DockerImageName localstackImage = DockerImageName.parse("localstack/localstack");
-
-    @Rule
-    static LocalStackContainer localstack = new LocalStackContainer(localstackImage)
-            .withServices(S3);
-    static S3Client s3Client;
+    @Inject
+    S3Client s3Client;
+    @ConfigProperty(name = "bucket.name")
+    String bucketName;
 
     static Product product;
 
     @BeforeAll
     @Transactional
     static void startLocalStack(){
-        localstack.start();
-
-        System.out.println("getEndpoint--------------:" + localstack.getEndpoint());
-        System.out.println("getAccessKey--------------:" + localstack.getAccessKey());
-        System.out.println("getRegion--------------:" + localstack.getRegion());
-        System.out.println("getSecretKey--------------:" + localstack.getSecretKey());
-        s3Client = S3Client
-                .builder()
-                .endpointOverride(localstack.getEndpoint())
-                .credentialsProvider(
-                        StaticCredentialsProvider.create(
-                                AwsBasicCredentials.create(localstack.getAccessKey(), localstack.getSecretKey())
-                        )
-                )
-                .region(Region.of(localstack.getRegion()))
-                .build();
-
-        CreateBucketRequest bucketRequest = CreateBucketRequest.builder()
-                .bucket("ecomarkets")
-                .build();
-
-        s3Client.createBucket(bucketRequest);
-
         product = new ProductBuilder().
                 name("Tomate").
                 description("Bolo de Banana Fitness (Zero Glúten e Lactose)").
@@ -73,9 +40,14 @@ public class ProductImageTest {
 
         product.persist();
     }
-
     @Test
     public void testS3Resource() {
+        CreateBucketRequest bucketRequest = CreateBucketRequest.builder()
+                .bucket(bucketName)
+                .build();
+
+        s3Client.createBucket(bucketRequest);
+
         File fileToUpload = new File("src/test/resources/ecomarkets/domain/core/product/acerola.jpg");
 
         given()
@@ -83,15 +55,14 @@ public class ProductImageTest {
             .when()
             .put("/api/product/%d/image".formatted(product.id))
             .then()
-            .statusCode(200); // Adjust the expected status code as needed
+            .statusCode(HttpStatus.SC_OK); // Adjust the expected status code as needed
 
         ListObjectsRequest listRequest = ListObjectsRequest.builder().bucket("ecomarkets").build();
 
-        //HEAD S3 objects to get metadata
         List<S3Object> r = s3Client.listObjects(listRequest).contents().stream()
                 .collect(Collectors.toList());
         r.size();
 
-        }
+    }
 
 }
