--- ecomarkets/src/test/resources/application.properties
@@ -1,16 +1,0 @@
-quarkus.datasource.devservices.port=3336
-quarkus.lambda.mock-event-server.dev-port=9090
-quarkus.lambda.mock-event-server.test-port=9091
-
-quarkus.hibernate-orm.database.generation=update
-quarkus.datasource.devservices.volumes."/local/test/data"=/var/lib/mysql
-
-quarkus.s3.endpoint-override=${LOCALSTACK_HOST}
-
-quarkus.s3.aws.region=us-east-1
-quarkus.s3.aws.credentials.type=static
-quarkus.s3.aws.credentials.static-provider.access-key-id=test
-quarkus.s3.aws.credentials.static-provider.secret-access-key=test
-
-
-bucket.name=ecomarkets
